Allow specifying the socket family (default to AF_INET)
For an application that wishes to say provide a local
telnet server (for application backend debugging) it is
quite useful to be able to specify that the type of that
socket is AF_UNIX (instead of AF_INET) for security reaons
and such.

This change allows such specification to be made (and
defaults to the existing value when not provided).