Use the module/package namespaced logger
It is a bad habit to use the root logger in
packages because it makes it impossible for consumers
of a library to specify logging levels for just say
that library.

So this switches to using the right namespaced
logger (and not the root logger) in the various
places the root logger was used.